+++ src/main/java/com/_ithon/speeksee/domain/voicefeedback/practice/service/PracticeService.java
@@ -1,4 +1,4 @@
-package com._ithon.speeksee.domain.voicefeedback.streaming.service;
+package com._ithon.speeksee.domain.voicefeedback.practice.service;
 
 import java.util.List;
 
@@ -8,9 +8,9 @@ import org.springframework.transaction.annotation.Transactional;
 
 import com._ithon.speeksee.domain.member.entity.Member;
 import com._ithon.speeksee.domain.member.repository.MemberRepository;
-import com._ithon.speeksee.domain.voicefeedback.streaming.dto.response.PracticeResponse;
-import com._ithon.speeksee.domain.voicefeedback.streaming.entity.ScriptPractice;
-import com._ithon.speeksee.domain.voicefeedback.streaming.repository.ScriptPracticeRepository;
+import com._ithon.speeksee.domain.voicefeedback.practice.dto.response.PracticeResponse;
+import com._ithon.speeksee.domain.voicefeedback.practice.entity.ScriptPractice;
+import com._ithon.speeksee.domain.voicefeedback.practice.repository.ScriptPracticeRepository;
 import com._ithon.speeksee.global.infra.exception.entityException.MemberNotFoundException;
 import com._ithon.speeksee.global.infra.exception.entityException.PracticeNotFoundException;
