+++ src/main/java/com/_ithon/speeksee/domain/voicefeedback/statistics/dto/PracticeChartPoint.java
@@ -1,4 +1,4 @@
-package com._ithon.speeksee.domain.voicefeedback.streaming.dto.response;
+package com._ithon.speeksee.domain.voicefeedback.statistics.dto;
 
 import io.swagger.v3.oas.annotations.media.Schema;
